requirements.txt
asgiref==3.2.9
boto==2.49.0
boto3==1.14.22
botocore==1.17.22
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-debug-toolbar==2.2
django-filter==2.3.0
django-storages==1.9.1
django-taggit==1.3.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.1
jmespath==0.10.0
Markdown==3.2.2
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
